Move intlDateTimeFormat into its own file
