+++ src/app/components/Footer/index.tsx
@@ -6,7 +6,7 @@ import React, { memo } from 'react'
 import { useSelector } from 'react-redux'
 import { Trans, useTranslation } from 'react-i18next'
 import { selectIsOpen } from 'app/state/wallet/selectors'
-import { intlDateTimeFormat } from '../DateFormatter'
+import { intlDateTimeFormat } from '../DateFormatter/intlDateTimeFormat'
 import { backend } from 'vendors/backend'
 import { BackendAPIs } from 'config'
 import { MobileFooterNavigation } from '../MobileFooterNavigation'
